Better install, etc

+++ install.sh
@@ -10,6 +10,7 @@ dryrun=false
 while [ $# -ge 1 ]; do
   case "$1" in
     -d) dryrun=true ;;
+    -f) rm -rf ${DOT_OLD} ;;
   esac
   shift
 done
@@ -17,18 +18,18 @@ done
 $dryrun && echo "Performing Dry run !"
 
 do_mv () {
-  $dryrun && echo "mv $1\t\t-> $2" && return
-  mv $1 $2
+  echo "mv $1\t\t-> $2" 
+  $dryrun || mv $1 $2
 }
 
 do_ln () {
-  $dryrun && echo "ln $1\t\t-> $2" && return
-  ln -s $1 $2
+  echo "ln $1\t\t-> $2"
+  $dryrun || ln -s $1 $2
 }
 
 do_rm () {
-  $dryrun && echo "rm $1" && return
-  rm -rf $1
+  echo "rm $1" 
+  $dryrun || rm -rf $1
 }
 
 
@@ -36,22 +37,28 @@ prepare () {
   [ -d ${DOT_OLD} ] && ! $dryrun && echo "Stopping! ${DOT_OLD} exists !" && exit
   echo "Installing from ${DOT_ROOT}"
   #do_rm ${DOT_OLD}
-  ! $dryrun && mkdir -p ${DOT_OLD}
+  $dryrun || mkdir -p ${DOT_OLD}
 }
 
 link_file () {
   local src=$1 dst=$2
+  local move=false skip=false
 
   if [ -f "$dst" -o -d "$dst" -o -L "$dst" ]; then
-    [ "$(readlink $dst)" != "$src" ] && do_mv $dst ${DOT_OLD}/ || echo "Skipping $dst" 
+    move=true
+    if [ "$(readlink $dst)" == "$src" ]; then
+      skip=true
+      move=false
+    fi
   fi
 
-  do_ln $src $dst
+  $move && do_mv $dst ${DOT_OLD}/
+  $skip || do_ln $src $dst
 }
 
 prepare
 
-FILES=("bashrc" "bash_profile" "tmux.conf" "gitconfig" "vimrc" "vim")
+FILES=("bashrc" "bash_profile" "tmux.conf" "gitconfig" "gitignore" "vimrc" "vim")
 
 link_file "${DOT_ROOT}/bin" "${HOME}/bin"
